+++ app/database/migrations/2019_06_24_192144_remove_info_ops_source_realm_target_dominion_constraint.php
@@ -18,8 +18,10 @@ class RemoveInfoOpsSourceRealmTargetDominionConstraint extends Migration
                 ->after('data')
                 ->default(true);
 
-            $table->dropForeign('info_ops_source_realm_id_foreign');
-            $table->dropForeign('info_ops_target_dominion_id_foreign');
+            if (DB::getDriverName() !== 'sqlite') {
+                $table->dropForeign('info_ops_source_realm_id_foreign');
+                $table->dropForeign('info_ops_target_dominion_id_foreign');
+            }
 
             $table->dropUnique(['source_realm_id', 'target_dominion_id', 'type']);