+++ app/database/seeds/DevelopmentSeeder.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Seeder;
 use OpenDominion\Factories\DominionFactory;
 use OpenDominion\Factories\RealmFactory;
+use OpenDominion\Factories\RoundFactory;
 use OpenDominion\Models\Dominion;
 use OpenDominion\Models\Race;
 use OpenDominion\Models\Round;
@@ -40,9 +41,9 @@ class DevelopmentSeeder extends Seeder
             $this->createRealmAndDominion($user);
 
             $this->command->info(<<<INFO
-
 Done seeding data.
-A development round, user and dominion have been created for your convenience.
+
+A test round, user and dominion have been created for your convenience.
 You may login with email '{$user->email}' and password '{$this->userPassword}'.
 
 INFO
@@ -52,7 +53,7 @@ INFO
 
     protected function createUser(): User
     {
-        $this->command->info('Creating dev user');
+        $this->command->info('Creating test user');
 
         $user = User::create([
             'email' => 'email@example.com',
@@ -69,38 +70,41 @@ INFO
 
     protected function createRound(): Round
     {
-        $this->command->info('Creating development round');
+        $this->command->info('Creating test round');
 
-        $startDate = today();
+        $roundFactory = app(RoundFactory::class);
 
-        return Round::create([
-            'round_league_id' => RoundLeague::where('key', 'standard')->firstOrFail()->id,
-            'number' => 1,
-            'name' => 'Dev Round',
-            'start_date' => $startDate,
-            'end_date' => (clone $startDate)->addDays(49), // 50 minus today
-        ]);
+        return $roundFactory->create(
+            RoundLeague::firstOrFail(),
+            today(),
+            12,
+            6,
+            2,
+            true
+        );
     }
 
     protected function createRealmAndDominion(User $user): Dominion
     {
-        $this->command->info('Creating realm');
+        $this->command->info('Creating test realm');
 
         $realmFactory = app(RealmFactory::class);
 
         $round = Round::firstOrFail();
+
+        /** @var Race $humanRace */
         $humanRace = Race::where('name', 'Human')->firstOrFail();
 
         $realm = $realmFactory->create($round, $humanRace->alignment);
 
-        $this->command->info('Creating dominion');
+        $this->command->info('Creating test dominion');
 
         return $this->dominionFactory->create(
             $user,
             $realm,
             $humanRace,
-            'Test Ruler Name',
-            'Test Dominion Name'
+            'Developer',
+            'My Dominion'
         );
     }
 }