+++ src/Services/Dominion/Actions/ExploreActionService.php
@@ -4,13 +4,12 @@ namespace OpenDominion\Services\Dominion\Actions;
 
 use DB;
 use OpenDominion\Calculators\Dominion\Actions\ExplorationCalculator;
+use OpenDominion\Exceptions\GameException;
 use OpenDominion\Helpers\LandHelper;
 use OpenDominion\Models\Dominion;
 use OpenDominion\Services\Dominion\HistoryService;
 use OpenDominion\Services\Dominion\QueueService;
 use OpenDominion\Traits\DominionGuardsTrait;
-use RuntimeException;
-use Throwable;
 
 class ExploreActionService
 {
@@ -41,11 +40,16 @@ class ExploreActionService
      * @param Dominion $dominion
      * @param array $data
      * @return array
-     * @throws Throwable
+     * @throws GameException
      */
     public function explore(Dominion $dominion, array $data): array
     {
         $this->guardLockedDominion($dominion);
+
+        if($dominion->round->hasOffensiveActionsDisabled())
+        {
+            throw new GameException('Exploration has been disabled for the remainder of the round.');
+        }
 
         $data = array_only($data, array_map(function ($value) {
             return "land_{$value}";
@@ -56,13 +60,13 @@ class ExploreActionService
         $totalLandToExplore = array_sum($data);
 
         if ($totalLandToExplore === 0) {
-            throw new RuntimeException('Exploration was not begun due to bad input.');
+            throw new GameException('Exploration was not begun due to bad input.');
         }
 
         $maxAfford = $this->explorationCalculator->getMaxAfford($dominion);
 
         if ($totalLandToExplore > $maxAfford) {
-            throw new RuntimeException("You do not have enough platinum and/or draftees to explore for {$totalLandToExplore} acres.");
+            throw new GameException("You do not have enough platinum and/or draftees to explore for {$totalLandToExplore} acres.");
         }
 
         // todo: refactor. see training action service. same with other action services