+++ src/Services/Dominion/Actions/InvadeActionService.php
@@ -9,6 +9,7 @@ use OpenDominion\Calculators\Dominion\LandCalculator;
 use OpenDominion\Calculators\Dominion\MilitaryCalculator;
 use OpenDominion\Calculators\Dominion\RangeCalculator;
 use OpenDominion\Calculators\Dominion\SpellCalculator;
+use OpenDominion\Exceptions\GameException;
 use OpenDominion\Models\Dominion;
 use OpenDominion\Models\GameEvent;
 use OpenDominion\Models\Unit;
@@ -17,8 +18,6 @@ use OpenDominion\Services\Dominion\ProtectionService;
 use OpenDominion\Services\Dominion\QueueService;
 use OpenDominion\Services\NotificationService;
 use OpenDominion\Traits\DominionGuardsTrait;
-use RuntimeException;
-use Throwable;
 
 class InvadeActionService
 {
@@ -173,7 +172,7 @@ class InvadeActionService
      * @param Dominion $target
      * @param array $units
      * @return array
-     * @throws Throwable
+     * @throws GameException
      */
     public function invade(Dominion $dominion, Dominion $target, array $units): array
     {
@@ -181,24 +180,29 @@ class InvadeActionService
             // Checks
             $this->guardLockedDominion($dominion);
 
+            if($dominion->round->hasOffensiveActionsDisabled())
+            {
+                throw new GameException('Invasions have been disabled for the remainder of the round.');
+            }
+
             if ($this->protectionService->isUnderProtection($dominion)) {
-                throw new RuntimeException('You cannot invade while under protection');
+                throw new GameException('You cannot invade while under protection');
             }
 
             if ($this->protectionService->isUnderProtection($target)) {
-                throw new RuntimeException('You cannot invade dominions which are under protection');
+                throw new GameException('You cannot invade dominions which are under protection');
             }
 
             if (!$this->rangeCalculator->isInRange($dominion, $target)) {
-                throw new RuntimeException('You cannot invade dominions outside of your range');
+                throw new GameException('You cannot invade dominions outside of your range');
             }
 
             if ($dominion->round->id !== $target->round->id) {
-                throw new RuntimeException('Nice try, but you cannot invade cross-round');
+                throw new GameException('Nice try, but you cannot invade cross-round');
             }
 
             if ($dominion->realm->id === $target->realm->id) {
-                throw new RuntimeException('Nice try, but you cannot invade your realmies');
+                throw new GameException('Nice try, but you cannot invade your realmies');
             }
 
             // Sanitize input
@@ -206,31 +210,31 @@ class InvadeActionService
             $landRatio = $this->rangeCalculator->getDominionRange($dominion, $target) / 100;
 
             if (!$this->hasAnyOP($dominion, $units)) {
-                throw new RuntimeException('You need to send at least some units');
+                throw new GameException('You need to send at least some units');
             }
 
             if (!$this->allUnitsHaveOP($dominion, $units)) {
-                throw new RuntimeException('You cannot send units that have no OP');
+                throw new GameException('You cannot send units that have no OP');
             }
 
             if (!$this->hasEnoughUnitsAtHome($dominion, $units)) {
-                throw new RuntimeException('You don\'t have enough units at home to send this many units');
+                throw new GameException('You don\'t have enough units at home to send this many units');
             }
 
             if (!$this->hasEnoughBoats($dominion, $units)) {
-                throw new RuntimeException('You do not have enough boats to send this many units');
+                throw new GameException('You do not have enough boats to send this many units');
             }
 
             if ($dominion->morale < static::MIN_MORALE) {
-                throw new RuntimeException('You do not have enough morale to invade others');
+                throw new GameException('You do not have enough morale to invade others');
             }
 
-            if (!$this->passes33PercentRule($dominion, $target, null, $units)) {
-                throw new RuntimeException('You need to leave more DP units at home, based on the OP you\'re sending out (33% rule)');
+            if (!$this->passes33PercentRule($dominion, $target, $units)) {
+                throw new GameException('You need to leave more DP units at home, based on the OP you\'re sending out (33% rule)');
             }
 
             if (!$this->passes54RatioRule($dominion, $target, $landRatio, $units)) {
-                throw new RuntimeException('You are sending out too much OP, based on your new home DP (5:4 rule)');
+                throw new GameException('You are sending out too much OP, based on your new home DP (5:4 rule)');
             }
 
             // Handle invasion results
@@ -251,6 +255,14 @@ class InvadeActionService
             $this->handleLandGrabs($dominion, $target);
 
             $this->invasionResult['attacker']['unitsSent'] = $units;
+
+            // Stat changes
+            // todo: move to own method
+            if ($this->invasionResult['result']['success']) {
+                $dominion->increment('stat_attacking_success');
+            } else {
+                $target->increment('stat_defending_success');
+            }
 
             // todo: move to GameEventService
             $this->invasionEvent = GameEvent::create([
@@ -335,7 +347,6 @@ class InvadeActionService
      * @param Dominion $dominion
      * @param Dominion $target
      * @param array $units
-     * @throws Throwable
      */
     protected function handlePrestigeChanges(Dominion $dominion, Dominion $target, array $units): void
     {
@@ -351,7 +362,7 @@ class InvadeActionService
 
         } elseif ($isInvasionSuccessful && ($range >= 75)) {
             $attackerPrestigeChange = (int)round(min(
-                (($target->prestige * (static::PRESTIGE_CHANGE_PERCENTAGE / 100)) + static::PRESTIGE_CHANGE_ADD), // Gained through invading
+                (($target->prestige * (($range / 100) / 12)) + static::PRESTIGE_CHANGE_ADD), // Gained through invading
                 (($dominion->prestige * (static::PRESTIGE_CAP_PERCENTAGE / 100)) + static::PRESTIGE_CHANGE_ADD) // But capped by depending on your current prestige
             ));
             $targetPrestigeChange = (int)round(($target->prestige * -(static::PRESTIGE_CHANGE_PERCENTAGE / 100)));
@@ -603,7 +614,6 @@ class InvadeActionService
      *
      * @param Dominion $dominion
      * @param Dominion $target
-     * @throws Throwable
      */
     protected function handleLandGrabs(Dominion $dominion, Dominion $target): void
     {
@@ -931,7 +941,6 @@ class InvadeActionService
      * @param Dominion $dominion
      * @param array $units
      * @param array $convertedUnits
-     * @throws Throwable
      */
     protected function handleReturningUnits(Dominion $dominion, array $units, array $convertedUnits): void
     {
@@ -967,7 +976,6 @@ class InvadeActionService
      * @param Dominion $dominion
      * @param Dominion $target
      * @param array $units
-     * @throws Throwable
      */
     protected function handleBoats(Dominion $dominion, Dominion $target, array $units): void
     {
@@ -1174,12 +1182,12 @@ class InvadeActionService
      *
      * @param Dominion $dominion
      * @param Dominion $target
-     * @param float $landRatio
      * @param array $units
      * @return bool
      */
-    protected function passes33PercentRule(Dominion $dominion, Dominion $target, ?float $landRatio, array $units): bool
+    protected function passes33PercentRule(Dominion $dominion, Dominion $target, array $units): bool
     {
+        $attackingForceOP = $this->militaryCalculator->getOffensivePower($dominion, $target->race->name, null, $units);
         $attackingForceDP = $this->militaryCalculator->getDefensivePower($dominion, null, null, $units);
         $currentHomeForcesDP = $this->militaryCalculator->getDefensivePower($dominion);